requirements.txt
# Core dependencies
streamlit>=1.24.0
pandas>=1.5.0
altair>=4.2.0
python-dateutil>=2.8.2

# Database
sqlite3

# AI and ML components
scikit-learn>=1.2.0
google-generativeai>=0.3.0

# API key management
python-dotenv>=0.21.0

# Terminal styling
colorama>=0.4.6